GHI-274 merge confirm password and set password
