+++ SafeUIKit/SafeAppUI/OnboardingFlow/MasterPasswordFlow/MasterPassword.storyboard
@@ -10,10 +10,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Set Password View Controller-->
+        <!--Password View Controller-->
         <scene sceneID="dnt-fH-pMc">
             <objects>
-                <viewController storyboardIdentifier="SetPasswordViewController" id="96a-oq-EKx" customClass="SetPasswordViewController" customModule="SafeAppUI" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="PasswordViewController" id="96a-oq-EKx" customClass="PasswordViewController" customModule="SafeAppUI" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="wXm-qa-Pl9" customClass="BackgroundImageView" customModule="SafeUIKit">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -92,50 +92,6 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="hpl-lU-zra" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="362" y="164"/>
-        </scene>
-        <!--Confirm Pasword View Controller-->
-        <scene sceneID="Vqq-zW-70W">
-            <objects>
-                <viewController storyboardIdentifier="ConfirmPaswordViewController" id="Gnv-1t-OLB" customClass="ConfirmPaswordViewController" customModule="SafeAppUI" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="yE1-xo-ade">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Repeat your master password" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="v2D-jj-zS3" customClass="H1Label" customModule="SafeUIKit">
-                                <rect key="frame" x="20" y="104" width="335" height="86"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="36"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="k0n-QG-KkA" customClass="VerifiableInput" customModule="SafeUIKit">
-                                <rect key="frame" x="20" y="210" width="335" height="150"/>
-                                <color key="backgroundColor" white="0.66666666669999997" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="150" placeholder="YES" id="Yh2-cs-5lO"/>
-                                </constraints>
-                            </view>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="k0n-QG-KkA" firstAttribute="top" secondItem="v2D-jj-zS3" secondAttribute="bottom" constant="20" id="6Ic-kc-6Q1"/>
-                            <constraint firstItem="o1b-OT-TXJ" firstAttribute="trailing" secondItem="v2D-jj-zS3" secondAttribute="trailing" constant="20" id="ANg-Tq-yVk"/>
-                            <constraint firstItem="k0n-QG-KkA" firstAttribute="trailing" secondItem="v2D-jj-zS3" secondAttribute="trailing" id="Ajw-Nr-5Xi"/>
-                            <constraint firstItem="v2D-jj-zS3" firstAttribute="leading" secondItem="o1b-OT-TXJ" secondAttribute="leading" constant="20" id="C6X-2Y-3Nf"/>
-                            <constraint firstItem="k0n-QG-KkA" firstAttribute="leading" secondItem="v2D-jj-zS3" secondAttribute="leading" id="Clq-X2-zxz"/>
-                            <constraint firstItem="v2D-jj-zS3" firstAttribute="top" secondItem="o1b-OT-TXJ" secondAttribute="top" constant="40" id="SkN-PF-Rv0"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="o1b-OT-TXJ"/>
-                    </view>
-                    <navigationItem key="navigationItem" id="bQe-ow-KpL"/>
-                    <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" prompted="NO"/>
-                    <connections>
-                        <outlet property="headerLabel" destination="v2D-jj-zS3" id="2i6-72-Bzq"/>
-                        <outlet property="verifiableInput" destination="k0n-QG-KkA" id="zW1-Ge-ehq"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="3b6-sU-7ff" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="1110" y="159"/>
         </scene>
         <!--Start View Controller-->
         <scene sceneID="dtT-0g-Knb">
