--- SafeUIKit/SafeAppUI/OnboardingFlow/MasterPasswordFlow/SetPasswordViewControllerTests.swift
@@ -1,46 +1,0 @@
-//
-//  Copyright © 2018 Gnosis Ltd. All rights reserved.
-//
-
-import XCTest
-@testable import SafeAppUI
-import SafeUIKit
-
-class SetPasswordViewControllerTests: XCTestCase {
-
-    // swiftlint:disable:next weak_delegate
-    let delegate = MockSetPasswordViewControllerDelegate()
-    var vc: SetPasswordViewController!
-
-    override func setUp() {
-        super.setUp()
-        delegate.wasCalled = false
-        vc = SetPasswordViewController.create(delegate: delegate)
-        vc.loadViewIfNeeded()
-    }
-
-    func test_whenLoaded_thenHasAllElements() {
-        XCTAssertNotNil(vc.headerLabel)
-        XCTAssertNotNil(vc.verifiableInput)
-    }
-
-    func test_whenLoaded_thenTextInputIsSecure() {
-        XCTAssertTrue(vc.verifiableInput.isSecure)
-    }
-
-    func test_whenPasswordSet_thenDelegateCalled() {
-        vc.verifiableInputDidReturn(vc.verifiableInput)
-        XCTAssertTrue(delegate.wasCalled)
-    }
-
-}
-
-class MockSetPasswordViewControllerDelegate: SetPasswordViewControllerDelegate {
-
-    var wasCalled = false
-
-    func didSetPassword(_ password: String) {
-        wasCalled = true
-    }
-
-}
